fix: address matgren's code review feedback
- Fix markdown rendering regression by guarding resolveOptionDisplay early-return
  with a check that resolvedDisplays has entries before short-circuiting
- Cap pageSize at 200 in options.ts to prevent unbounded queries
- Extract ~460 lines of utility code from CustomDataSection.tsx into
  packages/ui/src/backend/utils/customFieldRelationDisplay.ts
- Add AbortController to fetch cleanup in useEffect instead of boolean flag
- Replace empty catch blocks with console.debug logging per OM convention
- Add tests for customer_person_profile, catalog_product_variant,
  markdown regression, API error fallback, multi-value relations, and pageSize cap
