+++ packages/ui/src/backend/__tests__/CustomDataSection.test.tsx
@@ -88,7 +88,7 @@ describe('CustomDataSection relation display', () => {
         title="Custom fields"
         loadFields={async () => ({ fields, definitions })}
         labels={{
-          loading: 'Loading custom data…',
+          loading: 'Loading custom data\u2026',
           emptyValue: 'No value',
           noFields: 'No fields',
           saveShortcut: 'Save now',
@@ -157,7 +157,7 @@ describe('CustomDataSection relation display', () => {
         title="Custom fields"
         loadFields={async () => ({ fields, definitions })}
         labels={{
-          loading: 'Loading custom data…',
+          loading: 'Loading custom data\u2026',
           emptyValue: 'No value',
           noFields: 'No fields',
           saveShortcut: 'Save now',
@@ -172,5 +172,294 @@ describe('CustomDataSection relation display', () => {
       'href',
       `/backend/customers/people-v2/${encodeURIComponent(relationId)}`,
     )
+  })
+
+  it('renders text/richtext fields as markdown even when relation displays are empty', async () => {
+    ;(window as WindowWithOriginalFetch).__omOriginalFetch = jest.fn(async () => {
+      return new Response(
+        JSON.stringify({ items: [] }),
+        { status: 200, headers: { 'content-type': 'application/json' } },
+      )
+    })
+
+    const fields: CrudField[] = [
+      {
+        id: 'cf_notes',
+        label: 'Notes',
+        type: 'richtext',
+      },
+      {
+        id: 'cf_linked_case',
+        label: 'Linked case',
+        type: 'select',
+        options: [],
+        loadOptions: async () => [],
+      },
+    ]
+    const definitions: CustomFieldDefDto[] = [
+      {
+        key: 'notes',
+        kind: 'richtext',
+        label: 'Notes',
+      },
+      {
+        key: 'linked_case',
+        kind: 'relation',
+        label: 'Linked case',
+        optionsUrl: '/api/entities/relations/options?entityId=virtual%3Acase_study&labelField=title',
+      },
+    ]
+
+    renderWithProviders(
+      <CustomDataSection
+        entityId="customers:customer_entity"
+        values={{ cf_notes: '**Bold note**', cf_linked_case: relationId }}
+        onSubmit={async () => {}}
+        title="Custom fields"
+        loadFields={async () => ({ fields, definitions })}
+        labels={{
+          loading: 'Loading custom data\u2026',
+          emptyValue: 'No value',
+          noFields: 'No fields',
+          saveShortcut: 'Save now',
+          edit: 'Edit',
+          cancel: 'Cancel',
+        }}
+      />,
+    )
+
+    const notesContent = await screen.findByText('**Bold note**')
+    expect(notesContent).toBeInTheDocument()
+    expect(notesContent.className).toContain('text-sm')
+  })
+
+  it('resolves customer_person_profile with entity_id route context', async () => {
+    const entityId = 'aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee'
+    const profileId = 'pppppppp-1111-2222-3333-444444444444'
+
+    ;(window as WindowWithOriginalFetch).__omOriginalFetch = jest.fn(async () => {
+      return new Response(
+        JSON.stringify({
+          items: [
+            {
+              value: profileId,
+              label: 'John Doe Profile',
+              routeContext: { entity_id: entityId },
+            },
+          ],
+        }),
+        { status: 200, headers: { 'content-type': 'application/json' } },
+      )
+    })
+
+    const fields: CrudField[] = [
+      {
+        id: 'cf_person_profile',
+        label: 'Person Profile',
+        type: 'select',
+        options: [],
+        loadOptions: async () => [],
+      },
+    ]
+    const definitions: CustomFieldDefDto[] = [
+      {
+        key: 'person_profile',
+        kind: 'relation',
+        label: 'Person Profile',
+        optionsUrl: '/api/entities/relations/options?entityId=customers%3Acustomer_person_profile&labelField=title',
+      },
+    ]
+
+    renderWithProviders(
+      <CustomDataSection
+        entityId="customers:customer_entity"
+        values={{ cf_person_profile: profileId }}
+        onSubmit={async () => {}}
+        title="Custom fields"
+        loadFields={async () => ({ fields, definitions })}
+        labels={{
+          loading: 'Loading custom data\u2026',
+          emptyValue: 'No value',
+          noFields: 'No fields',
+          saveShortcut: 'Save now',
+          edit: 'Edit',
+          cancel: 'Cancel',
+        }}
+      />,
+    )
+
+    const link = await screen.findByRole('link', { name: 'John Doe Profile' })
+    expect(link).toHaveAttribute(
+      'href',
+      `/backend/customers/people-v2/${encodeURIComponent(entityId)}`,
+    )
+  })
+
+  it('resolves catalog_product_variant with product_id route context', async () => {
+    const productId = 'pppppppp-aaaa-bbbb-cccc-dddddddddddd'
+    const variantId = 'vvvvvvvv-1111-2222-3333-444444444444'
+
+    ;(window as WindowWithOriginalFetch).__omOriginalFetch = jest.fn(async () => {
+      return new Response(
+        JSON.stringify({
+          items: [
+            {
+              value: variantId,
+              label: 'Variant XL Red',
+              routeContext: { product_id: productId },
+            },
+          ],
+        }),
+        { status: 200, headers: { 'content-type': 'application/json' } },
+      )
+    })
+
+    const fields: CrudField[] = [
+      {
+        id: 'cf_variant_ref',
+        label: 'Product Variant',
+        type: 'select',
+        options: [],
+        loadOptions: async () => [],
+      },
+    ]
+    const definitions: CustomFieldDefDto[] = [
+      {
+        key: 'variant_ref',
+        kind: 'relation',
+        label: 'Product Variant',
+        optionsUrl: '/api/entities/relations/options?entityId=catalog%3Acatalog_product_variant&labelField=title',
+      },
+    ]
+
+    renderWithProviders(
+      <CustomDataSection
+        entityId="customers:customer_entity"
+        values={{ cf_variant_ref: variantId }}
+        onSubmit={async () => {}}
+        title="Custom fields"
+        loadFields={async () => ({ fields, definitions })}
+        labels={{
+          loading: 'Loading custom data\u2026',
+          emptyValue: 'No value',
+          noFields: 'No fields',
+          saveShortcut: 'Save now',
+          edit: 'Edit',
+          cancel: 'Cancel',
+        }}
+      />,
+    )
+
+    const link = await screen.findByRole('link', { name: 'Variant XL Red' })
+    expect(link).toHaveAttribute(
+      'href',
+      `/backend/catalog/products/${encodeURIComponent(productId)}/variants/${encodeURIComponent(variantId)}`,
+    )
+  })
+
+  it('falls back gracefully on API error', async () => {
+    ;(window as WindowWithOriginalFetch).__omOriginalFetch = jest.fn(async () => {
+      return new Response(
+        JSON.stringify({ error: 'Forbidden' }),
+        { status: 403, headers: { 'content-type': 'application/json' } },
+      )
+    })
+
+    const fields: CrudField[] = [
+      {
+        id: 'cf_some_relation',
+        label: 'Some Relation',
+        type: 'select',
+        options: [],
+        loadOptions: async () => [],
+      },
+    ]
+    const definitions: CustomFieldDefDto[] = [
+      {
+        key: 'some_relation',
+        kind: 'relation',
+        label: 'Some Relation',
+        optionsUrl: '/api/entities/relations/options?entityId=virtual%3Aexample&labelField=title',
+      },
+    ]
+
+    renderWithProviders(
+      <CustomDataSection
+        entityId="customers:customer_entity"
+        values={{ cf_some_relation: relationId }}
+        onSubmit={async () => {}}
+        title="Custom fields"
+        loadFields={async () => ({ fields, definitions })}
+        labels={{
+          loading: 'Loading custom data\u2026',
+          emptyValue: 'No value',
+          noFields: 'No fields',
+          saveShortcut: 'Save now',
+          edit: 'Edit',
+          cancel: 'Cancel',
+        }}
+      />,
+    )
+
+    const fallbackText = await screen.findByText(relationId)
+    expect(fallbackText).toBeInTheDocument()
+  })
+
+  it('resolves multi-value relation fields (array of UUIDs)', async () => {
+    const id1 = 'aaaaaaaa-1111-2222-3333-444444444444'
+    const id2 = 'bbbbbbbb-1111-2222-3333-444444444444'
+
+    ;(window as WindowWithOriginalFetch).__omOriginalFetch = jest.fn(async () => {
+      return new Response(
+        JSON.stringify({
+          items: [
+            { value: id1, label: 'Record Alpha' },
+            { value: id2, label: 'Record Beta' },
+          ],
+        }),
+        { status: 200, headers: { 'content-type': 'application/json' } },
+      )
+    })
+
+    const fields: CrudField[] = [
+      {
+        id: 'cf_multi_relation',
+        label: 'Multi Relation',
+        type: 'multiselect',
+        options: [],
+        loadOptions: async () => [],
+      },
+    ]
+    const definitions: CustomFieldDefDto[] = [
+      {
+        key: 'multi_relation',
+        kind: 'relation',
+        label: 'Multi Relation',
+        optionsUrl: '/api/entities/relations/options?entityId=virtual%3Aexample&labelField=title',
+      },
+    ]
+
+    renderWithProviders(
+      <CustomDataSection
+        entityId="customers:customer_entity"
+        values={{ cf_multi_relation: [id1, id2] }}
+        onSubmit={async () => {}}
+        title="Custom fields"
+        loadFields={async () => ({ fields, definitions })}
+        labels={{
+          loading: 'Loading custom data\u2026',
+          emptyValue: 'No value',
+          noFields: 'No fields',
+          saveShortcut: 'Save now',
+          edit: 'Edit',
+          cancel: 'Cancel',
+        }}
+      />,
+    )
+
+    const linkAlpha = await screen.findByRole('link', { name: 'Record Alpha' })
+    expect(linkAlpha).toBeInTheDocument()
+    const linkBeta = await screen.findByRole('link', { name: 'Record Beta' })
+    expect(linkBeta).toBeInTheDocument()
   })
 })
